+++ src/common_util.c
@@ -52,7 +52,7 @@ void copy_int_as_le(void *dst, void *src, size_t len) {
     }
 }
 
-int validate_path_component(const char *cmpnt, size_t len_s) {
+int validate_path_component(const char *cmpnt, uint8_t len_s) {
     for (uint8_t i = 0; i < len_s; i++) {
         unsigned char c = cmpnt[i];
         if (c & 0x80) {
